pyproject.toml
[project]
name = "psw"
version = "0.0.0"
description = "Build scripts for PSW"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "contourpy==1.3.1 ; python_full_version >= '3.10'",
    "cycler==0.12.1 ; python_full_version >= '3.8'",
    "fonttools==4.55.0 ; python_full_version >= '3.8'",
    "icalendar==6.1.0 ; python_full_version >= '3.8'",
    "kiwisolver==1.4.7 ; python_full_version >= '3.8'",
    "matplotlib==3.9.2 ; python_full_version >= '3.9'",
    "mistletoe==1.4.0 ; python_full_version >= '3.5' and python_full_version < '4'",
    "numpy==2.1.3 ; python_full_version >= '3.10'",
    "packaging==24.2 ; python_full_version >= '3.8'",
    "pillow==11.0.0 ; python_full_version >= '3.9'",
    "pyparsing==3.2.0 ; python_full_version >= '3.9'",
    "python-dateutil==2.9.0.post0 ; python_full_version >= '2.7'",
    "pyyaml==6.0.2 ; python_full_version >= '3.8'",
    "requests>=2.32.4",
    "six==1.16.0 ; python_full_version >= '2.7'",
    "tzdata==2024.2 ; python_full_version >= '2'",
    "unidecode==1.3.8 ; python_full_version >= '3.5'",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
]

[tool.pytest]
pythonpath = ["src"]
testpaths = ["src/test"]
